# docs/codex/rituvia-production-2026-07-23/05_API_CONTRACTS.openapi.yaml
openapi: 3.1.0
info:
  title: RITUVIA Production API
  version: 1.0.0-draft
  description: Server-authoritative API contract. Adapt paths to the existing repository without weakening invariants.
servers:
  - url: https://staging.example.invalid/api
security:
  - cookieSession: []
...
paths:
...
  /v1/numerology/calculate:
    post:
      tags: [Readings]
      security: []
      operationId: calculateNumerology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [birthDate]
              properties:
                birthDate: {type: string, format: date}
                latinName: {type: string, maxLength: 120}
      responses:
        '200': {description: Deterministic results and visible calculation steps}
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: __Host-rituvia_session
...